read-text-dirs: add option for reading labels from subdirectories
Added an option for specifying a root directory so that its subdirectories could become labels (equivalent for listing all of them on the command line).

This is similar to mallet's [import-dir](http://mallet.cs.umass.edu/import.php).  This is a convenience feature, particularly applicable when the number of labels is large.